ARM: Exynos4: add support for GATE BLOCK clocks
EXYNOS4_CLKGATE_BLOCK register can be used to disable the respective
multimedia block hardware planes. It acts similar to the power domains.
This patch adds transparent support for this method of gating of
multimedia blocks. New clocks are added as parents to the respective
block bus clocks. This patch has been tested on NURI and UniversalC210
boards, which have bootloader which disable all multimedia blocks with
both GATE BLOCK method and power domains.

Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>