ARM: SAMSUNG: Move samsung-time to drivers/clocksource
This patch moves the Samsung PWM-based high resolution timer support
code from arch/arm/plat-samsung to drivers/clocksource.

This is a prerequisite for further work on making the driver more
multiplatform and Device Tree friendly.

In addition, the driver is renamed to samsung_pwm, since it uses PWM
timers available on Samsung SoCs for timekeeping purposes.

Signed-off-by: Tomasz Figa <t.figa@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>